@@ -4,12 +4,12 @@ version: 1.0.0
 # homepage: https://www.example.com
 
 environment:
-  sdk: ">=3.0.5 <4.0.0"
+  sdk: '>=3.3.2 <4.0.0'
 
 dependencies:
-  stormberry: ^0.13.0
+  stormberry: ^0.14.0
 
 dev_dependencies:
   build_runner: ^2.3.3
-  lints: ^2.0.0
+  lints: ^3.0.0
   test: ^1.16.0
