@@ -4,7 +4,7 @@ version: 1.0.0
 publish_to: none
 
 environment:
-  sdk: ">=3.0.5 <4.0.0"
+  sdk: '>=3.3.2 <4.0.0'
 
 dependency_overrides:
   analyzer: ^5.3.1
@@ -13,13 +13,13 @@ dependencies:
   db:
     path: ../db
 
-  freezed_annotation: ^2.2.0
-  json_annotation: ^4.7.0
+  freezed_annotation: ^2.4.1
+  json_annotation: ^4.8.0
 
 dev_dependencies:
   build_runner: ^2.3.3
-  freezed: ^2.3.2
-  json_serializable: ^6.5.4
-  lints: ^2.0.0
+  freezed: ^2.4.7
+  json_serializable: ^6.7.1
+  lints: ^3.0.0
   test: ^1.16.0
 
