@@ -4,7 +4,7 @@ version: 1.0.0+1
 publish_to: none
 
 environment:
-  sdk: ">=3.0.5 <4.0.0"
+  sdk: '>=3.3.2 <4.0.0'
 
 dependencies:
   # hashing
@@ -28,9 +28,9 @@ dependencies:
     path: packages/shared
 
   # ORM
-  stormberry: ^0.13.0
+  stormberry: ^0.14.0
 
 dev_dependencies:
-  mocktail: ^0.3.0
+  mocktail: ^1.0.3
   test: ^1.19.2
-  very_good_analysis: ^3.1.0
+  very_good_analysis: ^5.1.0
