@@ -19,7 +19,7 @@ part 'widgets/_body.dart';
 part 'widgets/_tile.dart';
 
 class ChatScreen extends StatelessWidget {
-  const ChatScreen({Key? key}) : super(key: key);
+  const ChatScreen({super.key});
 
   @override
   Widget build(BuildContext context) {
