@@ -40,7 +40,7 @@ part 'widgets/_cover_photo.dart';
 part 'widgets/_profile_photo.dart';
 
 class EditProfileScreen extends StatelessWidget {
-  const EditProfileScreen({Key? key}) : super(key: key);
+  const EditProfileScreen({super.key});
 
   @override
   Widget build(BuildContext context) {
