@@ -21,7 +21,7 @@ part 'widgets/_message_input.dart';
 part '_state.dart';
 
 class InboxScreen extends StatelessWidget {
-  const InboxScreen({Key? key}) : super(key: key);
+  const InboxScreen({super.key});
 
   @override
   Widget build(BuildContext context) {
