@@ -30,7 +30,7 @@ part '_auth_listener.dart';
 part '_follow_listener.dart';
 
 class ProfileScreen extends StatelessWidget {
-  const ProfileScreen({Key? key}) : super(key: key);
+  const ProfileScreen({super.key});
 
   @override
   Widget build(BuildContext context) {
