@@ -74,9 +74,9 @@ class WidgetAnimator extends StatelessWidget {
   final Widget child;
 
   const WidgetAnimator({
-    Key? key,
+    super.key,
     required this.child,
-  }) : super(key: key);
+  });
 
   @override
   Widget build(BuildContext context) {
