@@ -2,7 +2,7 @@
 
 part of 'post.dart';
 
-extension PostRepositories on Database {
+extension PostRepositories on Session {
   PostRepository get posts => PostRepository._(this);
 }
 
@@ -12,7 +12,7 @@ abstract class PostRepository
         KeyedModelRepositoryInsert<PostInsertRequest>,
         ModelRepositoryUpdate<PostUpdateRequest>,
         ModelRepositoryDelete<int> {
-  factory PostRepository._(Database db) = _PostRepository;
+  factory PostRepository._(Session db) = _PostRepository;
 
   Future<PostView?> queryPost(int id);
   Future<List<PostView>> queryPosts([QueryParams? params]);
@@ -40,11 +40,12 @@ class _PostRepository extends BaseRepository
   Future<List<int>> insert(List<PostInsertRequest> requests) async {
     if (requests.isEmpty) return [];
     var values = QueryValues();
-    var rows = await db.query(
-      'INSERT INTO "posts" ( "uid", "caption", "has_image", "image_url", "has_video", "video_url", "likes", "comments", "created_at" )\n'
-      'VALUES ${requests.map((r) => '( ${values.add(r.uid)}:int8, ${values.add(r.caption)}:text, ${values.add(r.hasImage)}:boolean, ${values.add(r.imageUrl)}:text, ${values.add(r.hasVideo)}:boolean, ${values.add(r.videoUrl)}:text, ${values.add(r.likes)}:_int8, ${values.add(r.comments)}:_int8, ${values.add(r.createdAt)}:timestamp )').join(', ')}\n'
-      'RETURNING "id"',
-      values.values,
+    var rows = await db.execute(
+      Sql.named(
+          'INSERT INTO "posts" ( "uid", "caption", "has_image", "image_url", "has_video", "video_url", "likes", "comments", "created_at" )\n'
+          'VALUES ${requests.map((r) => '( ${values.add(r.uid)}:int8, ${values.add(r.caption)}:text, ${values.add(r.hasImage)}:boolean, ${values.add(r.imageUrl)}:text, ${values.add(r.hasVideo)}:boolean, ${values.add(r.videoUrl)}:text, ${values.add(r.likes)}:_int8, ${values.add(r.comments)}:_int8, ${values.add(r.createdAt)}:timestamp )').join(', ')}\n'
+          'RETURNING "id"'),
+      parameters: values.values,
     );
     var result = rows.map<int>((r) => TextEncoder.i.decode(r.toColumnMap()['id'])).toList();
 
@@ -55,13 +56,13 @@ class _PostRepository extends BaseRepository
   Future<void> update(List<PostUpdateRequest> requests) async {
     if (requests.isEmpty) return;
     var values = QueryValues();
-    await db.query(
-      'UPDATE "posts"\n'
-      'SET "uid" = COALESCE(UPDATED."uid", "posts"."uid"), "caption" = COALESCE(UPDATED."caption", "posts"."caption"), "has_image" = COALESCE(UPDATED."has_image", "posts"."has_image"), "image_url" = COALESCE(UPDATED."image_url", "posts"."image_url"), "has_video" = COALESCE(UPDATED."has_video", "posts"."has_video"), "video_url" = COALESCE(UPDATED."video_url", "posts"."video_url"), "likes" = COALESCE(UPDATED."likes", "posts"."likes"), "comments" = COALESCE(UPDATED."comments", "posts"."comments"), "created_at" = COALESCE(UPDATED."created_at", "posts"."created_at")\n'
-      'FROM ( VALUES ${requests.map((r) => '( ${values.add(r.id)}:int8::int8, ${values.add(r.uid)}:int8::int8, ${values.add(r.caption)}:text::text, ${values.add(r.hasImage)}:boolean::boolean, ${values.add(r.imageUrl)}:text::text, ${values.add(r.hasVideo)}:boolean::boolean, ${values.add(r.videoUrl)}:text::text, ${values.add(r.likes)}:_int8::_int8, ${values.add(r.comments)}:_int8::_int8, ${values.add(r.createdAt)}:timestamp::timestamp )').join(', ')} )\n'
-      'AS UPDATED("id", "uid", "caption", "has_image", "image_url", "has_video", "video_url", "likes", "comments", "created_at")\n'
-      'WHERE "posts"."id" = UPDATED."id"',
-      values.values,
+    await db.execute(
+      Sql.named('UPDATE "posts"\n'
+          'SET "uid" = COALESCE(UPDATED."uid", "posts"."uid"), "caption" = COALESCE(UPDATED."caption", "posts"."caption"), "has_image" = COALESCE(UPDATED."has_image", "posts"."has_image"), "image_url" = COALESCE(UPDATED."image_url", "posts"."image_url"), "has_video" = COALESCE(UPDATED."has_video", "posts"."has_video"), "video_url" = COALESCE(UPDATED."video_url", "posts"."video_url"), "likes" = COALESCE(UPDATED."likes", "posts"."likes"), "comments" = COALESCE(UPDATED."comments", "posts"."comments"), "created_at" = COALESCE(UPDATED."created_at", "posts"."created_at")\n'
+          'FROM ( VALUES ${requests.map((r) => '( ${values.add(r.id)}:int8::int8, ${values.add(r.uid)}:int8::int8, ${values.add(r.caption)}:text::text, ${values.add(r.hasImage)}:boolean::boolean, ${values.add(r.imageUrl)}:text::text, ${values.add(r.hasVideo)}:boolean::boolean, ${values.add(r.videoUrl)}:text::text, ${values.add(r.likes)}:_int8::_int8, ${values.add(r.comments)}:_int8::_int8, ${values.add(r.createdAt)}:timestamp::timestamp )').join(', ')} )\n'
+          'AS UPDATED("id", "uid", "caption", "has_image", "image_url", "has_video", "video_url", "likes", "comments", "created_at")\n'
+          'WHERE "posts"."id" = UPDATED."id"'),
+      parameters: values.values,
     );
   }
 }
