@@ -2,7 +2,7 @@
 
 part of 'user.dart';
 
-extension UserRepositories on Database {
+extension UserRepositories on Session {
   UserRepository get users => UserRepository._(this);
 }
 
@@ -12,7 +12,7 @@ abstract class UserRepository
         KeyedModelRepositoryInsert<UserInsertRequest>,
         ModelRepositoryUpdate<UserUpdateRequest>,
         ModelRepositoryDelete<int> {
-  factory UserRepository._(Database db) = _UserRepository;
+  factory UserRepository._(Session db) = _UserRepository;
 
   Future<UserView?> queryUser(int id);
   Future<List<UserView>> queryUsers([QueryParams? params]);
@@ -40,11 +40,12 @@ class _UserRepository extends BaseRepository
   Future<List<int>> insert(List<UserInsertRequest> requests) async {
     if (requests.isEmpty) return [];
     var values = QueryValues();
-    var rows = await db.query(
-      'INSERT INTO "users" ( "first_name", "followers", "following", "posts", "stories", "last_name", "username", "email", "password", "bio", "birthday", "image_url", "cover_url" )\n'
-      'VALUES ${requests.map((r) => '( ${values.add(r.firstName)}:text, ${values.add(r.followers)}:_int8, ${values.add(r.following)}:_int8, ${values.add(r.posts)}:_int8, ${values.add(r.stories)}:_int8, ${values.add(r.lastName)}:text, ${values.add(r.username)}:text, ${values.add(r.email)}:text, ${values.add(r.password)}:text, ${values.add(r.bio)}:text, ${values.add(r.birthday)}:timestamp, ${values.add(r.imageURL)}:text, ${values.add(r.coverURL)}:text )').join(', ')}\n'
-      'RETURNING "id"',
-      values.values,
+    var rows = await db.execute(
+      Sql.named(
+          'INSERT INTO "users" ( "first_name", "followers", "following", "posts", "stories", "last_name", "username", "email", "password", "bio", "birthday", "image_url", "cover_url" )\n'
+          'VALUES ${requests.map((r) => '( ${values.add(r.firstName)}:text, ${values.add(r.followers)}:_int8, ${values.add(r.following)}:_int8, ${values.add(r.posts)}:_int8, ${values.add(r.stories)}:_int8, ${values.add(r.lastName)}:text, ${values.add(r.username)}:text, ${values.add(r.email)}:text, ${values.add(r.password)}:text, ${values.add(r.bio)}:text, ${values.add(r.birthday)}:timestamp, ${values.add(r.imageURL)}:text, ${values.add(r.coverURL)}:text )').join(', ')}\n'
+          'RETURNING "id"'),
+      parameters: values.values,
     );
     var result = rows.map<int>((r) => TextEncoder.i.decode(r.toColumnMap()['id'])).toList();
 
@@ -55,13 +56,13 @@ class _UserRepository extends BaseRepository
   Future<void> update(List<UserUpdateRequest> requests) async {
     if (requests.isEmpty) return;
     var values = QueryValues();
-    await db.query(
-      'UPDATE "users"\n'
-      'SET "first_name" = COALESCE(UPDATED."first_name", "users"."first_name"), "followers" = COALESCE(UPDATED."followers", "users"."followers"), "following" = COALESCE(UPDATED."following", "users"."following"), "posts" = COALESCE(UPDATED."posts", "users"."posts"), "stories" = COALESCE(UPDATED."stories", "users"."stories"), "last_name" = COALESCE(UPDATED."last_name", "users"."last_name"), "username" = COALESCE(UPDATED."username", "users"."username"), "email" = COALESCE(UPDATED."email", "users"."email"), "password" = COALESCE(UPDATED."password", "users"."password"), "bio" = COALESCE(UPDATED."bio", "users"."bio"), "birthday" = COALESCE(UPDATED."birthday", "users"."birthday"), "image_url" = COALESCE(UPDATED."image_url", "users"."image_url"), "cover_url" = COALESCE(UPDATED."cover_url", "users"."cover_url")\n'
-      'FROM ( VALUES ${requests.map((r) => '( ${values.add(r.id)}:int8::int8, ${values.add(r.firstName)}:text::text, ${values.add(r.followers)}:_int8::_int8, ${values.add(r.following)}:_int8::_int8, ${values.add(r.posts)}:_int8::_int8, ${values.add(r.stories)}:_int8::_int8, ${values.add(r.lastName)}:text::text, ${values.add(r.username)}:text::text, ${values.add(r.email)}:text::text, ${values.add(r.password)}:text::text, ${values.add(r.bio)}:text::text, ${values.add(r.birthday)}:timestamp::timestamp, ${values.add(r.imageURL)}:text::text, ${values.add(r.coverURL)}:text::text )').join(', ')} )\n'
-      'AS UPDATED("id", "first_name", "followers", "following", "posts", "stories", "last_name", "username", "email", "password", "bio", "birthday", "image_url", "cover_url")\n'
-      'WHERE "users"."id" = UPDATED."id"',
-      values.values,
+    await db.execute(
+      Sql.named('UPDATE "users"\n'
+          'SET "first_name" = COALESCE(UPDATED."first_name", "users"."first_name"), "followers" = COALESCE(UPDATED."followers", "users"."followers"), "following" = COALESCE(UPDATED."following", "users"."following"), "posts" = COALESCE(UPDATED."posts", "users"."posts"), "stories" = COALESCE(UPDATED."stories", "users"."stories"), "last_name" = COALESCE(UPDATED."last_name", "users"."last_name"), "username" = COALESCE(UPDATED."username", "users"."username"), "email" = COALESCE(UPDATED."email", "users"."email"), "password" = COALESCE(UPDATED."password", "users"."password"), "bio" = COALESCE(UPDATED."bio", "users"."bio"), "birthday" = COALESCE(UPDATED."birthday", "users"."birthday"), "image_url" = COALESCE(UPDATED."image_url", "users"."image_url"), "cover_url" = COALESCE(UPDATED."cover_url", "users"."cover_url")\n'
+          'FROM ( VALUES ${requests.map((r) => '( ${values.add(r.id)}:int8::int8, ${values.add(r.firstName)}:text::text, ${values.add(r.followers)}:_int8::_int8, ${values.add(r.following)}:_int8::_int8, ${values.add(r.posts)}:_int8::_int8, ${values.add(r.stories)}:_int8::_int8, ${values.add(r.lastName)}:text::text, ${values.add(r.username)}:text::text, ${values.add(r.email)}:text::text, ${values.add(r.password)}:text::text, ${values.add(r.bio)}:text::text, ${values.add(r.birthday)}:timestamp::timestamp, ${values.add(r.imageURL)}:text::text, ${values.add(r.coverURL)}:text::text )').join(', ')} )\n'
+          'AS UPDATED("id", "first_name", "followers", "following", "posts", "stories", "last_name", "username", "email", "password", "bio", "birthday", "image_url", "cover_url")\n'
+          'WHERE "users"."id" = UPDATED."id"'),
+      parameters: values.values,
     );
   }
 }
